requirements.txt
mongoengine==0.27.0
multipledispatch==0.6.0
PyInquirer==1.0.3
pymongo==4.3.3
PyYAML==6.0
pyzmq==25.0.1
rospy==1.15.15
setuptools==65.6.3
strictyaml==1.7.3
transitions==0.9.0
watchdog==2.3.1
